Create README.md
Analyzing Data Using Python 


The goal of this exercise is analyzing data using Python. Some tasks done in the PyBank folder : 

Calculating profit and losses

Changes in profit and losses 

Max and Min Changes.

 The PyPoll folder contains analysis done on elections data and determining the winner of the elections. 


References
Chan, J. (2017). Learn Python in one day and learn it well : Python for beginners with hands-on project : the only book you need to start coding in Python immediately.